pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "read-no-evil-mcp"
version = "0.1.0"
description = "A secure email gateway MCP server that protects AI agents from prompt injection attacks in emails"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "read-no-evil-mcp contributors" }
]
keywords = ["mcp", "email", "security", "prompt-injection", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Communications :: Email",
]
dependencies = [
    "fastmcp",
    "transformers>=4.30",
    "torch",
    "structlog>=24",
    "imap-tools",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pytest",
    "pytest-asyncio",
]

[project.scripts]
read-no-evil-mcp = "read_no_evil_mcp.server:main"

[project.urls]
Homepage = "https://github.com/thekie/read-no-evil-mcp"
Repository = "https://github.com/thekie/read-no-evil-mcp"
Issues = "https://github.com/thekie/read-no-evil-mcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/read_no_evil_mcp"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "imap_tools.*",
    "mcp.*",
    "fastmcp.*",
    "transformers.*",
    "optimum.*",
    "pydantic_settings.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: marks tests as integration tests (may be slower, require model loading)",
]
